requirements.txt
aiohttp==0.22.0a0 \
    --hash=sha256:ea8e1dd3eb8b06ff15cdbecdd15db9eb54920546ba78dd09ec362381fbf5d8b1 \
    --hash=sha256:eea2b0b8a0957bc76bdbd1212794c474e5ccf67147fd115a62c2278d9e29579c \
    --hash=sha256:0b1fdce93dfc714c9667de93073f38931337d34877895138674770c7bf31b145 \
    --hash=sha256:31a59f8b76cde1ee89e26edcf9da3b89e0920024f97720806ef194ec5ebaed68 \
    --hash=sha256:5c6f1d5d62117d6962cf046a6490fa7798a2939629bf289694d3178ce6c03349
python-decouple==3.0 \
    --hash=sha256:99834c9ff7ce5c5b2f4a18bc0880753e54ea3aaabc3cfb16961a92d5046665df
multidict==1.1.0 \
    --hash=sha256:d0961c144e99387e6d531fd9354af10409ebeb8e5634ff610d16a5e3f5ba91c8 \
    --hash=sha256:5d9089267feab7a1cc12d2aa619754170405b43f809a8f2c19d45200e652d866 \
    --hash=sha256:a082f86dc0dbd65b46d72266902223e6627ffd3d50975cadca1a8ead9eb71b0a \
    --hash=sha256:45909b0ba17f8b595baf238e4929dd9643a7e71b7a2e4bf42803cad815a3f7af \
    --hash=sha256:c765e21e4bb54bc7b2d805d1a2704a1b7fc735e2d1136b8c7c99de95c4d0bc67 \
    --hash=sha256:044e915cec4871f30865ed9a8c44d845bd3f2f38f54932de2051322ee14ea561 \
    --hash=sha256:2d3ad25c6d9e849682ade28c591ee993a119172ce153026eb7596d442fc6947a \
    --hash=sha256:9748da96ce765cfb94f9d597a8c06a342fa1a181d5662ac248e7df06656552e8 \
    --hash=sha256:2f33750b9d0df41c83bc6a46931d218f0d0bad4d7799e6877be8ad33bdc64c19
statsd==3.2.1 \
    --hash=sha256:7aff40c6cdda703193fac6231410c6c0ad5939ce053a168016501e982c755c84 \
    --hash=sha256:3fa92bf0192af926f7a0d9be031fe3fd0fbaa1992d42cf2f07e68f76ac18288e
chardet==2.3.0 \
    --hash=sha256:e53e38b3a4afe6d1132de62b7400a4ac363452dc5dfcf8d88e8e0cce663c68aa
